requirements.txt
Django==4.2.5
djangorestframework==3.14.0
pip==23.2.1
gunicorn==21.2.0